refactor(fs): PATH 上の実行可能ファイル探索を fs::which へ切り出す
compiler::resolve が使っていた find_in_path/existing_executable は、
更新チェック機能 (#53) で cargo-install-update の有無判定にも使う
共通処理。偶然の重複ではないので、依存ゼロの共有層である fs へ移す。
